Refactoring `FlowControlException` out of FlowController
