+++ gax/src/test/java/com/google/api/gax/batching/v2/BatcherImplTest.java
@@ -36,9 +36,9 @@ import static com.google.common.truth.Truth.assertThat;
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutures;
 import com.google.api.core.SettableApiFuture;
+import com.google.api.gax.batching.v2.FlowControlException.MaxOutstandingElementCountReachedException;
+import com.google.api.gax.batching.v2.FlowControlException.MaxOutstandingRequestBytesReachedException;
 import com.google.api.gax.batching.v2.FlowControlSettings.LimitExceededBehavior;
-import com.google.api.gax.batching.v2.FlowController.MaxOutstandingElementCountReachedException;
-import com.google.api.gax.batching.v2.FlowController.MaxOutstandingRequestBytesReachedException;
 import com.google.api.gax.rpc.ApiCallContext;
 import com.google.api.gax.rpc.UnaryCallable;
 import com.google.api.gax.rpc.testing.FakeBatchableApi.LabeledIntList;
